port

+++ kernel/sysproc.cc
@@ -71,7 +71,7 @@ sys_pause(void) {
         n = 0;
     acquire(&tickslock);
     ticks0 = ticks;
-    while (ticks - ticks0 < n) {
+    while (ticks - ticks0 < static_cast<uint>(n)) {
         if (killed(myproc())) {
             release(&tickslock);
             return -1;
